+++ src/select/base.rs
@@ -1,6 +1,4 @@
-use crate::parse_common::Position;
-use crate::tree::MdqNode;
-use crate::tree_ref::MdqNodeRef;
+use crate::select::SelectResult;
 
 pub trait Selector<'a, I: Copy> {
     fn matches(&self, item: I) -> bool;
@@ -13,26 +11,4 @@ pub trait Selector<'a, I: Copy> {
             None
         }
     }
-}
-
-pub enum SelectResult<'a> {
-    One(MdqNodeRef<'a>),
-    Multi(&'a Vec<MdqNode>),
-}
-
-pub type ParseResult<T> = Result<T, ParseErrorReason>;
-
-pub const SELECTOR_SEPARATOR: char = '|';
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ParseError {
-    pub position: Position,
-    pub reason: ParseErrorReason,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum ParseErrorReason {
-    UnexpectedCharacter(char),
-    UnexpectedEndOfInput,
-    InvalidSyntax(String),
 }
