+++ src/select/section.rs
@@ -1,8 +1,9 @@
 use crate::fmt_str::inlines_to_plain_string;
 use crate::matcher::Matcher;
 use crate::parsing_iter::ParsingIterator;
-use crate::select::base::{ParseResult, SelectResult, Selector};
+use crate::select::base::Selector;
 use crate::select::util::require_whitespace;
+use crate::select::{ParseResult, SelectResult};
 use crate::tree::Section;
 
 #[derive(Debug, PartialEq)]
